+++ src/__tests__/register.test.tsx
@@ -3,7 +3,7 @@ import fetchMock from 'jest-fetch-mock';
 import { signIn } from 'next-auth/react';
 import { randomUUID } from 'crypto';
 import RegisterPage from '@/pages/register';
-import { UpdatedResponse } from '@/features/common/commonTypes';
+import { UpdatedResponse } from '@/features/common/common-types';
 import { RegisterFormValues } from '@/features/account/components/register-form';
 import { queryWrapper } from '@/lib/test-utils';
