+++ src/features/manage/components/change-password-form.tsx
@@ -1,7 +1,7 @@
 import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { z } from 'zod';
-import { GetProfileResponse } from '@/features/manage/manageTypes';
+import { GetProfileResponse } from '@/features/manage/manage-types';
 import { Input } from '@/components/input';
 import { Button } from '@/components/button';
 import { ButtonGroup } from '@/components/button-group';
