+++ src/features/user/hooks/use-update-user.ts
@@ -1,7 +1,7 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { useSession } from 'next-auth/react';
-import { UpdatedResponse } from '@/features/common/commonTypes';
-import { UpdateUserRequest } from '@/features/user/userTypes';
+import { UpdatedResponse } from '@/features/common/common-types';
+import { UpdateUserRequest } from '@/features/user/user-types';
 import { fetcher } from '@/lib/fetch';
 import { config } from '@/lib/config';
